% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_gains.R
\name{cumulative_gains}
\alias{cumulative_gains}
\title{Cumulative gains}
\usage{
cumulative_gains(x, prob_col = NA, truth_col = NA, positive = "Yes")
}
\arguments{
\item{x}{A data frame containing predicted probabilities of a target event and the actual outcome/class.}

\item{prob_col}{The unquoted name of the column with probabilities of the event of interest.}

\item{truth_col}{The unquoted name of the column with the actual outcome/class.}

\item{positive}{The value in `truth_col` that identifies the event of interest. Defaults to 'Yes'.}
}
\value{
A data frame with the following columns:

row        = row numbers \cr
pct        = percentiles \cr
prop_pop   = proportion of the population targeted (row / n) \cr
cum_events = cumulated number of actual positives captured \cr
gain       = proportion of all positives captured (cum_events / total positives) \cr
baseline   = expected gain from random targeting (equal to prop_pop)
}
\description{
Calculates a cumulative gains curve after sorting observations by descending
predicted probability. The gain at a given point is the proportion of all
actual positives (events) that have been captured by targeting the top rows.
}
\examples{
cumulative_gains(predictions,
   prob_col  = Yes,
   truth_col = Churn)
}
