% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Predictions from a customer churn model.}
\format{
A data frame with 2145 rows and 4 variables:
\describe{
  \item{predict}{Predicted class}
  \item{No}{Predicted probability of class 'No'}
  \item{Yes}{Predicted probability of class 'Yes'}
  \item{Churn}{Actual class}
  ...
}
}
\usage{
data
}
\description{
A dataset containing 2145 observations with four columns specifying predicted probabilities and predicted and actual class.
}
\keyword{datasets}
