% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelimpact-package.R
\docType{package}
\name{modelimpact-package}
\alias{modelimpact}
\alias{modelimpact-package}
\title{modelimpact: assess the business impact of churn prediction models}
\description{
Calculate the financial impact of using a churn model in terms of cost,
revenue, profit, ROI, gains, lift and threshold-based payoff, and visualise
the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PeerChristensen/modelimpact}
  \item Report bugs at \url{https://github.com/PeerChristensen/modelimpact/issues}
}

}
\keyword{internal}
