% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{modelimpact-plots}
\alias{modelimpact-plots}
\alias{autoplot.mi_profit}
\alias{autoplot.mi_cost_revenue}
\alias{autoplot.mi_roi}
\alias{autoplot.mi_gains}
\alias{autoplot.mi_lift}
\alias{autoplot.mi_marginal}
\alias{autoplot.mi_thresholds}
\alias{autoplot.mi_roc}
\alias{autoplot.mi_compare}
\alias{autoplot.mi_bootstrap}
\alias{plot_profit}
\alias{plot_cost_revenue}
\alias{plot_roi}
\alias{plot_gains}
\alias{plot_lift}
\alias{plot_marginal}
\alias{plot_thresholds}
\alias{plot_roc}
\title{Plot modelimpact results}
\usage{
autoplot.mi_profit(object, ...)

autoplot.mi_cost_revenue(object, ...)

autoplot.mi_roi(object, ...)

autoplot.mi_gains(object, ...)

autoplot.mi_lift(object, ...)

autoplot.mi_marginal(object, ...)

autoplot.mi_thresholds(object, ...)

autoplot.mi_roc(object, slope = NULL, ...)

autoplot.mi_compare(object, ...)

autoplot.mi_bootstrap(object, ...)

plot_profit(data)

plot_cost_revenue(data)

plot_roi(data)

plot_gains(data)

plot_lift(data)

plot_marginal(data)

plot_thresholds(data)

plot_roc(data, slope = NULL)
}
\arguments{
\item{object, data}{The data frame returned by the matching modelimpact
function (e.g. the output of [profit()] for `autoplot()` / `plot_profit()`).}

\item{...}{Unused, for S3 compatibility.}

\item{slope}{For ROC objects only: optional slope of an iso-profit line. When
supplied, the cost-sensitive optimal operating point (maximising
`tpr - slope * fpr`) is highlighted. The business-optimal slope equals
`(cost_fp / value_fn) * (n_neg / n_pos)`.}
}
\value{
A `ggplot` object.
}
\description{
Every analysis function in the package returns a classed data frame that can
be visualised directly with \code{\link[ggplot2]{autoplot}()}, for example
\code{autoplot(profit(...))}. The `plot_*()` functions are thin,
back-compatible wrappers around the corresponding `autoplot()` method.
\pkg{ggplot2} is only required when a plot is actually drawn.
}
