% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profit.R
\name{profit}
\alias{profit}
\title{Calculate profit}
\usage{
profit(
  x,
  fixed_cost = 0,
  var_cost = 0,
  tp_val = 0,
  prob_accept = 1,
  prob_col = NA,
  truth_col = NA,
  positive = "Yes"
)
}
\arguments{
\item{x}{A data frame containing predicted probabilities of a target event and the actual outcome/class.}

\item{fixed_cost}{Fixed cost (e.g. of a campaign)}

\item{var_cost}{Variable cost (e.g. discount offered)}

\item{tp_val}{The average value of a True Positive}

\item{prob_accept}{Probability of the offer being accepted. Variable cost is only incurred when accepted. Defaults to 1.}

\item{prob_col}{The unquoted name of the column with probabilities of the event of interest.}

\item{truth_col}{The unquoted name of the column with the actual outcome/class.}

\item{positive}{The value in `truth_col` that identifies the event of interest. Defaults to 'Yes'.}
}
\value{
A data frame with the following columns:

row       = row numbers \cr
pct       = percentiles \cr
profit    = profit for number of rows selected
}
\description{
Calculates profit after sorting observations.
}
\examples{
profit(predictions,
     fixed_cost = 1000,
     var_cost   = 100,
     tp_val     = 2000,
     prob_col   = Yes,
     truth_col  = Churn)

}
