% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_pr.R
\name{roc_pr}
\alias{roc_pr}
\title{ROC and precision-recall curve data}
\usage{
roc_pr(x, prob_col = NA, truth_col = NA, positive = "Yes")
}
\arguments{
\item{x}{A data frame containing predicted probabilities of a target event and the actual outcome/class.}

\item{prob_col}{The unquoted name of the column with probabilities of the event of interest.}

\item{truth_col}{The unquoted name of the column with the actual outcome/class.}

\item{positive}{The value in `truth_col` that identifies the event of interest. Defaults to 'Yes'.}
}
\value{
A data frame with the following columns:

threshold = probability at each operating point \cr
tp        = true positives when classifying every row at or above the point as positive \cr
fp        = false positives \cr
tpr       = true positive rate / recall / sensitivity (tp / all positives) \cr
fpr       = false positive rate (fp / all negatives) \cr
precision = tp / (tp + fp) \cr
recall    = same as tpr
}
\description{
Calculates the points of the ROC curve and the precision-recall curve by
walking down the observations sorted by descending predicted probability.
The returned data frame contains everything needed to draw both curves and to
overlay an iso-profit / cost-sensitive operating point.
}
\examples{
roc_pr(predictions,
   prob_col  = Yes,
   truth_col = Churn)
}
