% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{roi}
\alias{roi}
\title{Calculate Return on investment (ROI)}
\usage{
roi(x, fixed_cost = 0, var_cost = 0, tp_val = 0, prob_col = NA, truth_col = NA)
}
\arguments{
\item{x}{A data frame containing predicted probabilities of a target event and the actual outcome/class.}

\item{fixed_cost}{Fixed cost (e.g. of a campaign)}

\item{var_cost}{Variable cost (e.g. discount offered)}

\item{tp_val}{The average value of a True Positive}

\item{prob_col}{The unquoted name of the column with probabilities of the event of interest.}

\item{truth_col}{The unquoted name of the column with the actual outcome/class. Possible values are 'Yes' and 'No'.}
}
\value{
A data frame with the following columns:

row       = row numbers \cr
pct       = percentiles \cr
cum_rev   = cumulated revenue \cr
cost_sum  = cumulated costs \cr
roi       = return on investment
}
\description{
Calculates ROI after sorting observations with ROI defined as (Current Value - Start Value) / Start Value
}
\examples{
roi(predictions,
   fixed_cost = 1000,
   var_cost   = 100,
   tp_val     = 2000,
   prob_col   = Yes,
   truth_col  = Churn)
}
